+++ corehq/apps/hqcase/tests/test_dbaccessors.py
@@ -1,11 +1,11 @@
 from django.test import TestCase
 from casexml.apps.case.dbaccessors import get_open_case_docs_in_domain, \
-    get_open_case_ids_in_domain, get_number_of_cases_by_filters, \
-    get_all_case_owner_ids
+    get_open_case_ids_in_domain, get_number_of_cases_by_filters
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.hqcase.dbaccessors import get_number_of_cases_in_domain, \
     get_case_ids_in_domain, get_case_types_for_domain, get_cases_in_domain, \
-    get_case_ids_in_domain_by_owner, get_number_of_cases_in_domain_by_owner
+    get_case_ids_in_domain_by_owner, get_number_of_cases_in_domain_by_owner, \
+    get_all_case_owner_ids
 
 
 class DBAccessorsTest(TestCase):
