+++ corehq/ex-submodules/casexml/apps/case/dbaccessors/all_cases.py
@@ -1,21 +1,6 @@
 from dimagi.utils.couch.database import iter_docs
 
 
-def get_all_case_owner_ids(domain):
-    """
-    Get all owner ids that are assigned to cases in a domain.
-    """
-    from casexml.apps.case.models import CommCareCase
-    key = ["all owner", domain]
-    submitted = CommCareCase.get_db().view(
-        'case/all_cases',
-        group_level=3,
-        startkey=key,
-        endkey=key + [{}],
-    ).all()
-    return set([row['key'][2] for row in submitted])
-
-
 def get_open_case_ids_in_domain(domain, type=None, owner_id=None):
     from casexml.apps.case.models import CommCareCase
     if owner_id is not None and type is None:
